Changing ranking from float to int

+++ rank-forge/rank-forge-core/src/main/java/com/rankforge/core/interfaces/RankingAlgorithm.java
@@ -26,5 +26,5 @@ import com.rankforge.core.models.PlayerStats;
  * Date 26/10/24
  */
 public interface RankingAlgorithm {
-    double calculateRank(PlayerStats stats);
+    int calculateRank(PlayerStats stats);
 }
